backend/requirements.txt
# MT5 Dashboard Backend Dependencies

# Web Server & Socket.io
python-socketio[asyncio-client]>=5.9.0
aiohttp>=3.9.0
aiohttp-cors>=0.7.0

# Database
aiosqlite>=0.19.0

# Async support
asyncio-mqtt>=0.16.0

# Data processing
numpy>=1.24.0
pandas>=2.0.0

# Utilities
python-dateutil>=2.8.2
pytz>=2023.3

# For MT5 integration (optional - for live trading)
MetaTrader5>=5.0.45

# For daemon mode (Linux/Mac)
python-daemon>=3.0.0; sys_platform != 'win32'

# For process monitoring
psutil>=5.9.0